Langfuse_Guardrails/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langfuse>=2.0.0
guardrails-ai>=0.5.0
python-dotenv==1.0.0